[#471] random message

+++ robot/rover/science/include/Carousel.h
@@ -7,17 +7,18 @@
 
 #include "Stoppable.h"
 #include "Updatable.h"
+#include <cstdint>
 
 class Carousel : public Stoppable, public Updatable {
     protected:
-        int currentCuvette;
+        uint8_t currentCuvette;
 
     public:
         void home();
-        void goToCuvette(int cuvetteId);
+        void goToCuvette(uint8_t cuvetteId);
         void nextCuvette();
         void previousCuvette();
-        int getCurrentCuvette() const { return currentCuvette; }
+        uint8_t getCurrentCuvette() const { return currentCuvette; }
 
         virtual void eStop() override;
         virtual void update(unsigned long deltaMicroSeconds) override;
