+++ robot/rover/science/include/Fan.h
@@ -7,11 +7,12 @@
 
 #include "Updatable.h"
 #include "Stoppable.h"
+#include <cstdint>
 
 class Fan :  public Stoppable, public Updatable {
 
     protected:
-        int fanPWM;
+        uint8_t fanPWM;
     public:
         void turnOn();
         void turnOff();
