+++ robot/rover/science/include/Pump.h
@@ -8,11 +8,12 @@
 
 #include "Updatable.h"
 #include "Stoppable.h"
+#include <cstdint>
 
 class Pump : public Stoppable, public Updatable{
     protected:
         unsigned long pumpTimeElapsed;
-        int pumpPWM;
+        uint8_t pumpPWM;
     public:
         virtual void pump();
         virtual void backpump();
